requirements.txt
llama-index==0.9.48
openai==1.14.2
python-dotenv==1.0.1
tqdm==4.66.2
PyMuPDF==1.23.22
